+++ module-services/service-appmgr/data/MmiActionsParams.cpp
@@ -7,18 +7,38 @@ using namespace app::manager::actions;
 
 void MMICustomResultParams::addMessage(const MMIResultMessage &message)
 {
-    messages.push_back(message);
+    messages.emplace_back(message);
 }
 
-auto MMICustomResultParams::getMessage() -> std::vector<MMIResultMessage>
+auto MMICustomResultParams::getMessage() const -> std::vector<MMIResultMessage>
 {
-    return std::move(messages);
+    return messages;
 }
 
-auto MMICustomResultParams::getType() -> MMIType
+auto MMICustomResultParams::getMessageType() const noexcept -> MMIType
 {
     return type;
 }
+
+void MMICustomResultParams::accept(Visitor &v, std::string &displayMessage)
+{
+    v.visit(*this, displayMessage);
+};
+
+void MMINoneSpecifiedResult::accept(Visitor &v, std::string &displayMessage)
+{
+    v.visit(*this, displayMessage);
+};
+
+auto MMICallForwardingResult::getData() const -> std::tuple<std::string, std::string, std::string, std::string>
+{
+    return std::make_tuple(voice, fax, sync, async);
+}
+
+void MMICallForwardingResult::accept(Visitor &v, std::string &displayMessage)
+{
+    v.visit(*this, displayMessage);
+};
 
 MMIParams::MMIParams(std::string mmiData) : mmiData{std::move(mmiData)}
 {}
@@ -28,18 +48,17 @@ std::string MMIParams::getData() const
     return mmiData;
 }
 
-MMIResultParams::MMIResultParams(MMIResult result) : result(result)
+MMIResultParams::MMIResultParams(MMIResult result, std::shared_ptr<MMICustomResultParams> customResult)
+    : result(result), customResult(std::move(customResult))
 {}
 
-MMIResultParams::MMIResultParams(std::shared_ptr<IMMICustomResultParams> cResult) : customResult(cResult)
-{}
 
 MMIResultParams::MMIResult MMIResultParams::getData() const noexcept
 {
     return result;
 }
 
-auto MMIResultParams::getCustomMessage() const noexcept -> std::shared_ptr<IMMICustomResultParams>
+auto MMIResultParams::getCustomData() const noexcept -> std::shared_ptr<MMICustomResultParams>
 {
     return customResult;
 }